Work on switching to Tables and residuals
Tables are the recommended way for going ahead and reading FITS files with astropy, so switching to that. In addition, Tables are easier to work with, so adding columns and data is being implemented with that in mind.
Residuals is almost done, only unsure thing is about the nclus!=cluster_number conversion and if the second expression actually does calculate for the entire array, but overall is close to done.

Progress in conversion is happily increasing!