Work on including tfitlin
Not sure how to get the guess for the planar solving at the moment. lstsq's seems like it would make a good guess for the ODR, but not sure.

Also, it seems like a lot of the code can be cut if the guess is determined by numpy instead of manually.